Refactored JS and updated footer email
