+++ index.html
@@ -6,7 +6,7 @@
     <meta name="description" content="A form to allow students and alumni get tickets to a yearly CEAS Tribunal event called Luau.">
     <title>Luau Tickets</title>
 
-    <link rel="stylesheet" type="text/css" href="styles/style.css">
+    <link rel="stylesheet" type="text/css" href="build/style.min.css">
     <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.8.2/css/all.css">
 </head>
 <body>
@@ -56,9 +56,9 @@
 
     <footer>
         <p>University of Cincinnati | © 2019 College of Engineering and Applied Science Tribunal </p>
-        <p>For any questions or comments, contact us at careerfair@uc.edu.</p>      
+        <p>For any questions or comments, contact us at uc.eas.tribunal@gmail.com.</p>      
     </footer>
 
-    <script src="scripts/main.js" type=module></script>
+    <script src="build/main.min.js"></script>
 </body>
 </html>
